spdm-lite: handle chunked SET_CERTIFICATE requests
Reassemble plaintext CHUNK_SEND requests with LargeMessageCtx and dispatch SET_CERTIFICATE after the full request is buffered.

Do not support SET_CERTIFICATE inside secured messages yet. Reassembled secured SET_CERTIFICATE requests return an UnsupportedRequest error in the final CHUNK_SEND_ACK.

Move SET_CERTIFICATE tests into stack/src/tests with shared support helpers, and keep generic SPDM error imports ungated.
